@@ -1,4 +1,5 @@
 //imports
-import {SimpleTask, PluginError} from "justo-plugin";
-const child_process = require("child_process");
-const util = require("util");
+import {task} from "justo";
+
+//api
+export default task("org.justojs", "cli", require("./cli"));
